Ratio Converter and WPF Message Box bug
The ratio size was too big for my notebook with a lower resolution.
Given the fact that win 10 can even be executed on tablets, there should
be a ratio converter. The min size was maintained. Also, the message
boxes weren't being shown, because of the STA Thread not set. Fixed the
bug and added the enhancement.